refactor/constants: update to have common constants
Updated to have common constants.
