+++ src/app/reducers/network_state.js
@@ -3,7 +3,7 @@ import {
   NETWORK_DISCONNECTED,
   NETWORK_CONNECTED
 } from '../actions/network_state';
-import CONSTANTS from '../constants.json';
+import CONSTANTS from '../../constants.json';
 
 const initialState = {
   state: CONSTANTS.NETWORK_STATUS.CONNECTING
