+++ src/constants.json
@@ -1,4 +1,13 @@
 {
+  "DEFAULT_LIB_PATH": {
+    "win32": "./safe_authenticator.dll",
+    "darwin": "./libsafe_authenticator.dylib",
+    "linux": "./libsafe_authenticator.so"
+  },
+  "DEFAULT_CLIENT_HANDLE_KEYS": {
+    "UNAUTHORISED": "unauthorised",
+    "AUTHENTICATOR": "authenticator"
+  },
   "NETWORK_STATUS": {
     "CONNECTING": 0,
     "CONNECTED": 1,
