--- src/ffi/constants.json
@@ -1,16 +1,0 @@
-{
-  "DEFAULT_LIB_PATH": {
-    "win32": "./safe_authenticator.dll",
-    "darwin": "./libsafe_authenticator.dylib",
-    "linux": "./libsafe_authenticator.so"
-  },
-  "NETWORK_STATES": {
-    "CONNECTING": 0,
-    "CONNECTED": 1,
-    "DISCONNECTED": 2
-  },
-  "DEFAULT_CLIENT_HANDLE_KEYS": {
-    "UNAUTHORISED": "unauthorised",
-    "AUTHENTICATOR": "authenticator"
-  }
-}
